+++ CHANGELOG.md
@@ -12,6 +12,7 @@ with respect to its command line interface and HTTP interface.
 ### Added
 * Server: Single deployment rectification now waits for a reports a DeployState.
 * Server: Single deployment name cache harvests - hopefully will further speed deployments.
+* Server: pending state workaround for Singularity
 
 ### Fixed
 * Server: /deploy-queue-item now returns correct queue position and Resolution field.